Changes and additions to the project to automate execution.

+++ run_experiments2x.sh
@@ -4,10 +4,15 @@
 START_TIME=$(date +%s)
 echo "Start time: $(date '+%H:%M:%S')"
 
-# Apaga os outputs antigos, se existirem
-rm outputs/*
-#source speedupy_experiments_venv/bin/activate
+# Delete the old outputs / Deleta as saídas antigas
+rm *.txt
+if [ -d outputs ]; then
+    rm outputs/*
+else
+    mkdir outputs
+fi
 
+# Pre-configurations to run the Framework | Configurações prévias para executar o Framework
 pip install -r speedupy/requirements.txt
 
 chmod +x dnacc_prepare.sh
@@ -30,29 +35,30 @@ SOURCE_DIR="$ROOT_PATH/speedupy"
 
 # Define the list of destination paths / Define a lista de caminhos de destino
 
-# DESTINATIONS=("$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp03_quicksort/quicksort.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp02_look_and_say/look_and_say.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp11_gauss_legendre_quadrature/gauss_legendre_quadrature.py" "$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp04_heat_distribution_lu/heat_distribution_lu.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp08_fft/fft_speedupy.py" "$ROOT_PATH/speedupy_experiments/05msrgithubexps/05msrgithubexps_exp04_curves/curves_speedupy.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp14_pernicious_numbers/pernicious_numbers.py" "$ROOT_PATH/speedupy_experiments/05msrgithubexps/05msrgithubexps_exp02_cvar/cvar_speedupy.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp06_belief_propagation/belief_propagation_speedupy.py" "$ROOT_PATH/DNACC-with-speedupy/adapted_for_speedupy/examples/basic/basic_spheres.py" "$ROOT_PATH/DNACC-with-speedupy/adapted_for_speedupy/examples/walking_colloid/walking_colloid.py" "$ROOT_PATH/diversity-with-speedupy/Diversity_sims/vince_sim_speedupy.py" "$ROOT_PATH/Tiny-GSGP-with-speedupy/adapted_for_speedupy/TINY_GSHCGP.py" "$ROOT_PATH/epr-with-speedupy/analyse_speedupy.py" "$ROOT_PATH/qho-with-speedupy/qho2_speedupy.py")
+DESTINATIONS_0="$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp03_quicksort/quicksort.py" 
+DESTINATIONS_1="$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp02_look_and_say/look_and_say.py" 
+DESTINATIONS_2="$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp11_gauss_legendre_quadrature/gauss_legendre_quadrature.py" 
+DESTINATIONS_3="$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp04_heat_distribution_lu/heat_distribution_lu.py" 
+DESTINATIONS_4="$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp08_fft/fft_speedupy.py" 
+DESTINATIONS_5="$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp14_pernicious_numbers/pernicious_numbers.py" 
+DESTINATIONS_6="$ROOT_PATH/speedupy_experiments/05msrgithubexps/05msrgithubexps_exp02_cvar/cvar_speedupy.py" 
+DESTINATIONS_7="$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp06_belief_propagation/belief_propagation_speedupy.py" 
+DESTINATIONS_8="$ROOT_PATH/DNACC-with-speedupy/adapted_for_speedupy/examples/basic/basic_spheres.py" 
+DESTINATIONS_9="$ROOT_PATH/DNACC-with-speedupy/adapted_for_speedupy/examples/walking_colloid/walking_colloid.py" 
+DESTINATIONS_10="$ROOT_PATH/diversity-with-speedupy/Diversity_sims/vince_sim_speedupy.py" 
+DESTINATIONS_11="$ROOT_PATH/Tiny-GSGP-with-speedupy/adapted_for_speedupy/TINY_GSHCGP.py" 
+DESTINATIONS_12="$ROOT_PATH/epr-with-speedupy/analyse_speedupy.py"
+DESTINATIONS_13="$ROOT_PATH/qho-with-speedupy/qho2_speedupy.py"
+DESTINATIONS_14="$ROOT_PATH/speedupy_experiments/05msrgithubexps/05msrgithubexps_exp04_curves/curves_speedupy.py"
 
-DESTINATIONS=("$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp03_quicksort/quicksort.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp02_look_and_say/look_and_say.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp11_gauss_legendre_quadrature/gauss_legendre_quadrature.py" "$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp04_heat_distribution_lu/heat_distribution_lu.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp08_fft/fft_speedupy.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp14_pernicious_numbers/pernicious_numbers.py" "$ROOT_PATH/speedupy_experiments/05msrgithubexps/05msrgithubexps_exp02_cvar/cvar_speedupy.py" "$ROOT_PATH/speedupy_experiments/04benchproglangs/04benchpl_exp06_belief_propagation/belief_propagation_speedupy.py" "$ROOT_PATH/DNACC-with-speedupy/adapted_for_speedupy/examples/basic/basic_spheres.py" "$ROOT_PATH/DNACC-with-speedupy/adapted_for_speedupy/examples/walking_colloid/walking_colloid.py" "$ROOT_PATH/diversity-with-speedupy/Diversity_sims/vince_sim_speedupy.py" "$ROOT_PATH/Tiny-GSGP-with-speedupy/adapted_for_speedupy/TINY_GSHCGP.py" "$ROOT_PATH/epr-with-speedupy/analyse_speedupy.py")
-
-# DESTINATIONS=("$ROOT_PATH/speedupy_experiments/01pilots/01pilots_exp04_heat_distribution_lu/heat_distribution_lu.py")
+DESTINATIONS=($DESTINATIONS_0 $DESTINATIONS_1 $DESTINATIONS_2 $DESTINATIONS_3 $DESTINATIONS_4 $DESTINATIONS_5 $DESTINATIONS_6 $DESTINATIONS_7 $DESTINATIONS_8 $DESTINATIONS_9 $DESTINATIONS_10 $DESTINATIONS_11 $DESTINATIONS_12 $DESTINATIONS_13 $DESTINATIONS_14)
 
 # Define the list of arguments for each destination path / Define a lista de argumentos para cada caminho de destino
-# ARGUMENTS_0=("0.1" "0.05" "0.01" "0.005" "0.001")
-# ARGUMENTS_0=("1e1")
 ARGUMENTS_0=("1e1" "1e2" "1e3" "1e4" "1e5") # quicksort
 ARGUMENTS_1=("25" "30" "35" "40" "45") # look_and_say
 ARGUMENTS_2=("1000" "2000" "3000" "4000" "5000") # gauss_legendre_quadrature
 ARGUMENTS_3=("0.1" "0.05" "0.01" "0.005" "0.001") # heat_distribution_lu
 ARGUMENTS_4=("1000" "2000" "3000" "4000" "5000") # fft
-# ARGUMENTS_5=("1576862 -8567453 1648423 542312 512 -20135 1455678 52349" "4341212 -12312419 123123 5423672 107 20135 145678 52349" "-11124 -11124 62412 1412 107501 201635 15678 57849" "43441212 -22523123 6219 5143228 107501 20135 1455678 5234567849" "-111243412 -124122123 62192412 5281412 107501 201422635 123455678 5234567849") # curves
-# ARGUMENTS_6=("20000" "25000" "30000" "35000" "39000") # pernicious_numbers
-# ARGUMENTS_7=("1e6" "5e6" "10e6" "50e6" "100e6") # cvar
-# ARGUMENTS_8=("1000" "2000" "3000" "4000" "5000") # belief_propagation
-# ARGUMENTS_9=("40" "1000" "10000" "100000" "1000000") # dnacc_basic_spheres
-# ARGUMENTS_10=("-10" "-20" "-30" "-40" "-50") # dnacc_walking_colloid
-# ARGUMENTS_11=("1000000" "2000000" "3000000" "4000000" "5000000") # vince_sim
-# ARGUMENTS_12=("1" "3" "5" "7" "9") # TINY_GSHCGP
-# ARGUMENTS_13=("100" "200" "300" "400" "500") # analyse_epr
 ARGUMENTS_5=("20000" "25000" "30000" "35000" "39000") # pernicious_numbers
 ARGUMENTS_6=("1e6" "5e6" "10e6" "50e6" "100e6") # cvar
 ARGUMENTS_7=("1000" "2000" "3000" "4000" "5000") # belief_propagation
@@ -61,7 +67,15 @@ ARGUMENTS_9=("-10" "-20" "-30" "-40" "-50") # dnacc_walking_colloid
 ARGUMENTS_10=("1000000" "2000000" "3000000" "4000000" "5000000") # vince_sim
 ARGUMENTS_11=("1" "3" "5" "7" "9") # TINY_GSHCGP
 ARGUMENTS_12=("100" "200" "300" "400" "500") # analyse_epr
-# ARGUMENTS_14=("100" "500" "1000" "5000" "10000") # qho2
+ARGUMENTS_13=("100" "500" "1000" "5000" "6000") # qho2
+
+ARG_CURVES_1=("-11124 -11124 62412 1412 107501 201635 15678 57849")
+ARG_CURVES_2=("1576862 -8567453 1648423 542312 512 -20135 1455678 52349")
+ARG_CURVES_3=("4341212 -12312419 123123 5423672 107 20135 145678 52349")
+ARG_CURVES_4=("4341241 -1231219 1231423 5423672 10547 20135 145678 52349")
+ARG_CURVES_5=("-1112434 -1241223 6212412 5281412 107501 20142265 3455678 5467849")
+
+ARGUMENTS_14=("${ARG_CURVES_1}" "${ARG_CURVES_2}" "${ARG_CURVES_3}" "${ARG_CURVES_4}" "${ARG_CURVES_5}") # curves
 
 # Copy the source directory to each destination directory / Copia o diretório de origem para cada diretório de destino
 for i in "${!DESTINATIONS[@]}"; do
@@ -79,15 +93,16 @@ for i in "${!DESTINATIONS[@]}"; do
     DEST_DIR=$(dirname "$DEST")
     
     # Define the Python file to be executed
-    PYTHON_FILE="$DEST" 
-
+    PYTHON_FILE="$DEST"
+    
     # Define the arguments for the current destination
     ARGUMENTS_VAR="ARGUMENTS_$i[@]"
     ARGUMENTS=("${!ARGUMENTS_VAR}")
     
     # Define the output file name based on the Python file name
-    OUTPUT_FILE_NO_CACHE="$ROOT_PATH/$(basename "${PYTHON_FILE}_output_no_cache.txt")"
-    OUTPUT_FILE_MANUAL="$ROOT_PATH/intra_args_$(basename "${PYTHON_FILE}_output_manual.txt")"
+    OUTPUT_FILE_NO_CACHE=$ROOT_PATH/outputs/$(basename $PYTHON_FILE | cut -d. -f1)_output_no_cache.txt
+    # OUTPUT_FILE_MANUAL="$ROOT_PATH/intra_args_$(basename "${PYTHON_FILE}_output_manual.txt")"
+    OUTPUT_FILE_MANUAL="$ROOT_PATH/outputs/$(basename $PYTHON_FILE | cut -d. -f1)_output_spdpy_intra_args.txt"
 
     cd "$DEST_DIR"
     echo "Running $PYTHON_FILE with different arguments..."
@@ -102,15 +117,15 @@ for i in "${!DESTINATIONS[@]}"; do
         echo "-Execution mode: no-cache with $ARG"
 
         # Execute the Python script with the argument in 'no-cache' mode
-        for j in {1..2}; do
-            python3.12 $PYTHON_FILE $ARG --exec-mode no-cache | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_NO_CACHE
+        for j in {1..1}; do
+            python3.12 $PYTHON_FILE $(echo "$ARG") --exec-mode no-cache | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_NO_CACHE
         done
         
         echo "-Execution mode: manual with $ARG"
         
         # Execute the Python script with the argument in 'manual' mode
-        for j in {1..2}; do
-            python3.12 $PYTHON_FILE $ARG --exec-mode manual | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_MANUAL
+        for j in {1..1}; do
+            python3.12 $PYTHON_FILE $(echo "$ARG") --exec-mode manual | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_MANUAL
         done        
        # Delete the .speedupy folder after each argument / Deleta a pasta .speedupy após cada argumento
 		rm -rf "$DEST_DIR/.speedupy/"
@@ -125,14 +140,15 @@ for i in "${!DESTINATIONS[@]}"; do
     DEST_DIR=$(dirname "$DEST")
     
     # Define the Python file to be executed
-    PYTHON_FILE="$DEST" 
+    PYTHON_FILE="$DEST"
 
     # Define the arguments for the current destination
     ARGUMENTS_VAR="ARGUMENTS_$i[@]"
     ARGUMENTS=("${!ARGUMENTS_VAR}")
     
     # Define the output file name based on the Python file name    
-    OUTPUT_FILE_MANUAL="$ROOT_PATH/intra_exec_only_$(basename "${PYTHON_FILE}_output_manual.txt")"
+    # OUTPUT_FILE_MANUAL="$ROOT_PATH/intra_exec_only_$(basename "${PYTHON_FILE}_output_manual.txt")"
+    OUTPUT_FILE_MANUAL="$ROOT_PATH/outputs/$(basename $PYTHON_FILE | cut -d. -f1)_output_spdpy_intra_exec.txt"
 
     cd "$DEST_DIR"
     echo "Running $PYTHON_FILE with different arguments..."
@@ -143,15 +159,16 @@ for i in "${!DESTINATIONS[@]}"; do
         #Run setup.py before executing the Python script
         echo "Running setup.py for $PYTHON_FILE..."
         # python3.12 "$ROOT_PATH/speedupy/setup_exp/setup.py" "$PYTHON_FILE"
-        echo "-Execution mode: manual"
+        echo "-Execution mode: manual with $ARG"
         
         # Execute the Python script with the argument in 'manual' mode
-        for j in {1..2}; do
+        for j in {1..1}; do
             python3.12 "speedupy/setup_exp/setup.py" "$PYTHON_FILE"        
-            python3.12 $PYTHON_FILE $ARG --exec-mode manual | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_MANUAL
+            python3.12 $PYTHON_FILE $(echo "$ARG") --exec-mode manual | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_MANUAL
+            # Delete the .speedupy folder after each execution / Deleta a pasta .speedupy após cada execução
             rm -rf "$DEST_DIR/.speedupy/"
         done        
-       # Delete the .speedupy folder after each argument / Deleta a pasta .speedupy após cada argumento		
+       		
     done    
 done
 
@@ -163,41 +180,35 @@ for i in "${!DESTINATIONS[@]}"; do
     DEST_DIR=$(dirname "$DEST")
     
     # Define the Python file to be executed
-    PYTHON_FILE="$DEST" 
+    PYTHON_FILE="$DEST"
 
     # Define the arguments for the current destination
     ARGUMENTS_VAR="ARGUMENTS_$i[@]"
     ARGUMENTS=("${!ARGUMENTS_VAR}")
     
     # Define the output file name based on the Python file name    
-    OUTPUT_FILE_MANUAL="$ROOT_PATH/$(basename "intra_experiment_${PYTHON_FILE}_output_manual.txt")"
+    # OUTPUT_FILE_MANUAL="$ROOT_PATH/$(basename "intra_experiment_${PYTHON_FILE}_output_manual.txt")"
+    OUTPUT_FILE_MANUAL="$ROOT_PATH/outputs/$(basename $PYTHON_FILE | cut -d. -f1)_output_spdpy_intra_exp.txt"
 
     cd "$DEST_DIR"
     echo "Running $PYTHON_FILE with different arguments..."
+    
+    #Run setup.py before executing the Python script
+    echo "Running setup.py for $PYTHON_FILE..."
+    # python3.12 "$ROOT_PATH/speedupy/setup_exp/setup.py" "$PYTHON_FILE"
+    python3.12 "speedupy/setup_exp/setup.py" "$PYTHON_FILE"
+    echo "-Execution mode: manual with $ARG"
 
     # Run the script with each argument
-    for ARG in "${ARGUMENTS[@]}"; do
-    
-        #Run setup.py before executing the Python script
-        echo "Running setup.py for $PYTHON_FILE..."
-        # python3.12 "$ROOT_PATH/speedupy/setup_exp/setup.py" "$PYTHON_FILE"
-        python3.12 "speedupy/setup_exp/setup.py" "$PYTHON_FILE"
-        echo "-Execution mode: manual"
-        
+    for ARG in "${ARGUMENTS[@]}"; do    
         # Execute the Python script with the argument in 'manual' mode
-        for j in {1..2}; do
-            python3.12 $PYTHON_FILE $ARG --exec-mode manual | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_MANUAL
+        for j in {1..1}; do
+            python3.12 $PYTHON_FILE $(echo "$ARG") --exec-mode manual | tail -n 1 | cut -d':' -f2 >> $OUTPUT_FILE_MANUAL
         done        
-       # Delete the .speedupy folder after each argument / Deleta a pasta .speedupy após cada argumento
     done
     # Deleta os diretórios do speedupy copiados previamente
     rm -rf "$DEST_DIR/.speedupy/" "$DEST_DIR/speedupy/"
 done
-
-# Movendo saida para um diretório
-cd $ROOT_PATH
-mkdir outputs
-mv *.txt outputs/
 
 # Capture end time / Captura a hora de término
 END_TIME=$(date +%s)
